% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROICalc.R
\docType{package}
\name{ROICalc}
\alias{ROICalc}
\alias{ROICalc-package}
\title{ROICalc: A package with a simple ROI Calculator.}
\description{
This package is a calculator for determining the return
on an investment, based off of the intial investment,
share purchase and sell price, and the broker fees. An
example dataset is included highlighting a series of
ROI outputs at different sell values.
}
\references{
\url{https://www.investopedia.com/terms/r/returnoninvestment.asp}
}
\author{
Sean Sears, McMaster University \email{searssl@mcmaster.ca}
}
